Add test for controller

+++ src/main/java/com/wenting/repository/ItemRepository.java
@@ -3,6 +3,6 @@ package com.wenting.repository;
 import com.wenting.model.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItemRepository extends JpaRepository<Item, Long> {
-    Item findById(Long id);
+public interface ItemRepository extends JpaRepository<Item, Integer> {
+    Item findById(Integer id);
 }
